fix(imageview): do the pixel-buffer size math in size_t to avoid 32-bit overflow
imageview_create_ctx() allocated the pixel buffer with
kmalloc(iv->img_w * iv->img_h * 4). img_w and img_h are uint32_t, so the
multiplication was evaluated in 32-bit unsigned and only widened to
kmalloc's size_t argument afterwards — it could wrap before the conversion.

CodeQL flags this as cpp/integer-multiplication-cast-to-long (CWE-190).
Cast the first operand to size_t so the whole product is computed in 64-bit.
The current 512x384 pattern doesn't overflow, but the buffer size now scales
correctly and safely if the image dimensions ever grow.

+++ kernel/imageview_win.c
@@ -61,7 +61,11 @@ imageview_win_t* imageview_create_ctx(void) {
     iv->img_w = 512;
     iv->img_h = 384;
     iv->zoom = 1.0f;
-    iv->pixels = (uint8_t*)kmalloc(iv->img_w * iv->img_h * 4);
+    // Compute the allocation size in size_t: img_w/img_h are uint32_t, so
+    // img_w * img_h * 4 would otherwise be evaluated in 32-bit and could wrap
+    // before widening to kmalloc's size_t argument (CodeQL cpp/integer-
+    // multiplication-cast-to-long). Casting the first operand keeps it 64-bit.
+    iv->pixels = (uint8_t*)kmalloc((size_t)iv->img_w * iv->img_h * 4);
     if (iv->pixels) {
         generate_test_pattern(iv->pixels, iv->img_w, iv->img_h);
     }
